configured my startup class and program class
